Address SwiftLint warnings
