+++ Packages/EurofurenceKit/Sources/EurofurenceKit/Previewing/EurofurenceModel+Previewing.swift
@@ -303,119 +303,4 @@ extension EurofurenceModel {
         }
     }
     
-    // MARK: - Previewing Configuration
-    
-    private class PreviewingModelProperties: EurofurenceModelProperties {
-        
-        var containerDirectoryURL: URL {
-            FileManager.default.temporaryDirectory.appendingPathComponent("EurofurenceKitPreview")
-        }
-        
-        var synchronizationChangeToken: SynchronizationPayload.GenerationToken?
-        
-        func removeContainerResource(at url: URL) throws {
-            
-        }
-        
-        func proposedURL(forImageIdentifier identifier: String) -> URL {
-            let bundle = Bundle.previewing
-            let url = bundle.url(forResource: identifier, withExtension: nil)
-            
-            if let url = url {
-                return url
-            } else {
-                fatalError(
-                    "Requested preview URL for image not present in preview bundle. ID=\(identifier)"
-                )
-            }
-        }
-        
-    }
-    
-    private class PreviewingKeychain: Keychain {
-        
-        init(state: AuthenticationState) {
-            switch state {
-            case .authenticated:
-                credential = Credential(
-                    username: "Preview User",
-                    registrationNumber: 42,
-                    authenticationToken: AuthenticationToken("Unused"),
-                    tokenExpiryDate: .distantFuture
-                )
-                
-            case .unauthenticated:
-                credential = nil
-            }
-        }
-        
-        var credential: Credential?
-        
-    }
-    
-    private struct PreviewingEurofurenceAPI: EurofurenceAPI {
-        
-        var synchronizationPayload: SynchronizationPayload
-        
-        func fetchChanges(
-            since previousChangeToken: SynchronizationPayload.GenerationToken?
-        ) async throws -> SynchronizationPayload {
-            synchronizationPayload
-        }
-        
-        func downloadImage(
-            _ request: DownloadImage
-        ) async throws {
-            
-        }
-        
-        func requestAuthenticationToken(
-            using login: LoginRequest
-        ) async throws -> AuthenticatedUser {
-            AuthenticatedUser(
-                userIdentifier: 42,
-                username: "Previewing User",
-                token: "Unused",
-                tokenExpires: .distantFuture
-            )
-        }
-        
-        func registerPushNotificationToken(
-            registration: RegisterPushNotificationDeviceToken
-        ) async throws {
-            
-        }
-        
-        func requestLogout(
-            _ logout: LogoutRequest
-        ) async throws {
-            
-        }
-        
-        func fetchRemoteConfiguration() async -> RemoteConfiguration {
-            PreviewingRemoteConfiguration()
-        }
-        
-        func fetchMessages(for authenticationToken: AuthenticationToken) async throws -> [EurofurenceWebAPI.Message] {
-            []
-        }
-        
-        func markMessageAsRead(
-            request: AcknowledgeMessageRequest
-        ) async throws {
-            
-        }
-        
-    }
-    
-    private struct PreviewingRemoteConfiguration: RemoteConfiguration {
-        
-        let onChange = RemoteConfigurationChangedPublisher()
-        
-        subscript<Key>(key: Key.Type) -> Key.Value? where Key: RemoteConfigurationKey {
-            nil
-        }
-        
-    }
-    
 }
