+++ Packages/EurofurenceKit/Sources/EurofurenceKit/Previewing/Model Adapters/PreviewingRemoteConfiguration.swift
@@ -1,0 +1,11 @@
+import EurofurenceWebAPI
+
+struct PreviewingRemoteConfiguration: RemoteConfiguration {
+    
+    let onChange = RemoteConfigurationChangedPublisher()
+    
+    subscript<Key>(key: Key.Type) -> Key.Value? where Key: RemoteConfigurationKey {
+        nil
+    }
+    
+}
